OMAP3730: PM: Restrict PER pwrdm entering to OFF when DSS pwrdm is active
On 3730, it is observed that the PER power domain is entering
(atleast internally) to OFF state even when some of the sleep dependencies
are not met. This is impacting the DSS use cases as the GPIO bank 5 is
getting powered down and thus causing the display to be blanked even when
there is an active DMA happening to display frame buffer.

This is just a temporary work around while the issue is being
further investigated. This patch is an expensive one from an
idle power perspective as the lowest power state that PER can enter
is only RET if DSS is active.

Signed-off-by: Ranjith Lohithakshan <ranjithl@ti.com>