DM37xx: PM: Fix OPP table indexing
The OPP table indexing on DM37xx was using the min max VDD1/2
macros that is defined for 35xx. This was causing the cpufreq table inits
to be erroneous. This patch replaces the use of hard defined macros
with get_max_vdd1/2 functions which returns the correct bounds based on
cpu type.

Signed-off-by: Ranjith Lohithakshan <ranjithl@ti.com>