omap3: Remove auto-selection of PMICs
The current implementation almost assumes that only
TWL4030/TWL5030/TWl6030 are (or can be) used with the
OMAP processors. This is, however, not true.

This patch removes the automatic selection of the PMIC
from Kconfig. Current defaults are chosen in default
configuration omap2plus_defconfig and can easily be
added to configuration for any board, there should be
no compile/run-time impact.

Signed-off-by: Sanjeev Premi <premi@ti.com>