AM3517 : NAND :Fix for non-detection when application board is connected
The nand timing parameters as enumerated in nand_timings in
arch/arm/mach-omap2/board-flash.c are now no longer being used
in the device init for the Micron NAND chip.
Instead, no changes are now made to the timing parameters by
the kernel so that the u-boot assigned values are preserved.

Reviewed-by: Vaibhav Hiremath <hvaibhav@ti.com>
Signed-off-by: Abhilash K V <abhilash.kv@ti.com>
Signed-off-by: Sanjeev Premi <premi@ti.com>
Signed-off-by: Sriramakrishnan A G <srk@ti.com>