video: AM335X: Add support for LCD FB driver
LCD controller on AM335X SoC is an updated version of that found on DA850 SoC.
The da8xx-fb driver can support this LCD IP with some enhancements.

This patch adds support for updated interrupt configuration on the AM335X LCD IP.
1) Registers for setting and clearing interrupts are different.
2) Raw and masked status registers are different.

The updates have been tested on an emulation platform for new features and the
patch has been tested on DA850 platform to make sure nothing existing breaks.

Signed-off-by: Manjunathappa, Prakash <prakash.pm@ti.com>