ti81xx: mmc: unmount mmc/sd on card's hot removal
This patch is a rectification of commit id 605a4d1 which enabled
hot insertion/removal for mmc/sd cards in ti81xx.
It was found that though the mmc/sd card removal and insertion was
detected after the patch, the device was not unmounted on card
removal.
This patch fixes that problem.

Signed-off-by: Saxena, Parth <parth.saxena@ti.com>
Signed-off-by: satish <satish.patel@ti.com>