AM335x musb otg supporting adb and usb mouse
For AM335x the usb0 or usb1 can either act as
host or device mode. connect mini/micro-A plug for
host mode and mini/mincro-B plug for device mode to
the usb0/usb1 ports.
   usb0     usb1
   -------------
   device  device
   device  host
   host    device
   host    host
   -------------

Default behaviour:
usb0(port closer to the board leg) act as host(for mouse/keyboard)
and usb1 work in device mode(for ADB). usb1 will get enabled on boot. To
enable usb0 we will be inserting a dummy (Android Zero).

USB hot-plugging is not yet supported. So, In host mode mouse/keyboard
should be connected to the usbX port at the time of booting otherwise
run following command to wake it up.

echo F  >/proc/driver/musb_hdrc.X

Signed-off-by: Amit Pundir <amitpundir@ti.com>